hack for showing responses for a different user
There's no easy way to get a list of campaign ids for a user other than the logged in user,
so when a supervisor looks at a different user's responses, we query against the supervisor's
campaigns (which we assume are a superset of the participant's campaigns) and ignore the "invalid
user" server errors that are returned for campaigns the participant doesn't belong to.

This fixes the bug that was causing an error message to be displayed when supervisor's
tried to view students' responses.