Fix doc list parsing when user has no roles
Previously if user_roles was empty, parser was passing an empty string to RoleDocument.valueOf which caused an exception. Now the fromServerString() method handles unexpected values safely

This fixes the bug that was preventing documents that are linked to a user's class or campaign but not directly to the user from showing up in document list.