store document details in presenter instead of refetching
The old way didn't work for documents that were not in the user's personal docs. This fixes the bug where users can see a list of all documents associated with their class/campaign but clicking on one that didn't belong to them had no effect.

Gotcha: this assumes the doc list is already loaded. If user enters url for document detail manually, the document won't be loaded yet. App defaults to loading and showing list of all documents - this should be fixed to load all docs and then show detail.