clear form state when clearing fields
when editing, the document edit form stores a list of classes/campaigns originally associated with the document so it can compare those to the new list. previously, the list of originals was not being cleared when the form was cleared. this could cause a bug where, if you edited a form, then tried to create a form, classes or campaigns left over from the edit operation might be submitted with the request.

should be ok now.