ignore 0717 error when querying all campaigns
since api doesn't allow getting survey responses from multiple campaigns at once,
we do one query for each campaign and ignore errors that only make sense for
a specific campaign. we were already doing this for 0701 invalid user errors -
now we do it for 0717 analyst querying private campaign errors as well.

this fixes the bug where a user who was analyst or author of a private campaign
saw an error message when he first navigated to the responses tab.