fix history token problem
history management wasn't playing well with tab selection events, so disabled onSelected events and am just responding to direct clicks on tabs instead. this fixes the problem where a link from one tab to another always showed the default view of the new tab even if a different view was encoded in the history token. (was a problem with the dashboard links)

also got rid of a call to the campaign data api that was no longer needed in the responses tab since it was just being used to get the name and urn of campaigns, and those are available in the user info now.