put prompt responses in order
AndWellnessServer repo commit db73cf3d3d1da0fc0a39 added an index field to prompt response data.
This change sorts prompts by the index field when adding them to a survey response data object so that prompts will be displayed in the web gui in the same order they are displayed on the phone (which is the order in which they were defined in the xml config)

Note: this change will break the responses tab unless you are running revision number db73cf3d3d1da0fc0a39 or later of the server.