fix document bug
fixes a bug where, if user navigated to a document detail page and downloaded that document, then went to the detail page for a different document and clicked download again, he would get a copy of the same document.

fix was to attach one download click handler that calls a class method to get the document id rather than attaching a new event handler every time a document is loaded.