ACL for browse vs edit responses views
Browse and edit views both require that a campaign be selected.
In edit view, if user is a participant, he can see/edit his own responses only if the campaign is still running.
In edit view, if user is a supervisor, he can always see/edit responses by any member of the campaign.
In browse view, if user is an analyst, he can see responses by any members if the responses and the campaign are both shared.
In browse view, if user is an author or supervisor, he can see responses by any members of the campaign.

TODO: My Responses view that lets user view (but not edit) his own responses even when campaign is not running.