make supervisors option in campaign detail parsing
users who are not supervisors themselves cannot see supervisors and the supervisor field in the user_campaign_roles json would be omitted. checking for null prevents an error that was causing failure to load campaign detail page for users that are not supervisors of the campaign.

added a redirect to campaign list on failure so that if campaign detail does fail to load for some reason and user clicks on the campaign name a second time, you will see an error message again. (previously since the history token was already set to the campaign detail url, clicking on the campaign name again had no effect since gwt ignores redirects to the same history token and user would not see the error message as he continued clicking, which was confusing)