set google maps api key dynamically
google maps requires a different api key for every domain. this change checks for "andwellness" or "mobilizingcs" in the hostname of the machine serving the webpage and tries to choose the correct key

NOTE(2011/09/07): the key currently hard coded for mobilizingcs in this file doesn't work. someone needs to troubleshoot it or register another one