prompt new user to change password
github issues #151

When a new user tries to log in, instead of telling them to check name and password and try again, now the interface prompts the user to change his password. Note that right now the code is telling the difference between a real auth exception and a new account exception by parsing the error text because the same error code is returned for both. This needs to be updated after the back end is updated to return different error codes.