logout on auth exception when creating campaign
fixes AndWellnessFrontEnd issue #150

NOTE: campaign creation uses a formPanel instead of the AndWellnessDataService to submit the form because it involves uploading a file, so right now we're just parsing and checking the error response manually. If we use this front end for a while it's probably worth refactoring the data service a little bit so the parsing/error handling code can be called from outside the class and reused.