update userInfo when admin edits classes
This fixes the bug in front end issue #153 where an admin can create a class and add himself to it but then would not see the class in his classes tab until he reloads the app. The problem was that the class list in the classes tab comes from the userInfo object, which is loaded by MainApp when the app is loaded and only refreshed after that when an event triggers a refresh. Now the AdminClassEditPresenter fires an event every time a class is added/deleted/updated so MainApp knows to refresh userInfo.