style hack
Issue #160 - The last user or class in a search result is hidden because the list has height set to 100% of the dock layout panel's center container and the frozen header table row takes up one row's worth of that, so the last item of the list was getting pushed off the page. Hacky fix is to add an extra row at the end of the table. There is probably a correct a way to fix this...