upgrade to GWT 2.4 and recompile maps jar
We use a custom build of gwt-google-maps because the app needs access to some features that weren't offered in the wrapper online. (e.g., adding a widget to the infowindow)

This change updates the app to GWT 2.4 (by updating the gwt-user and gwt-dev jars), and also includes the custom maps jar recompiled with the GWT 2.4 libs. If a custom maps jar is still required the next time GWT is upgraded (that s, if there is still no official version of the gwt maps wrapper that supports all the methods we need), the maps jar may need to be recompiled again, which can be done using the src files in lib/gwt-google-maps-v3.tar.gz. (Unzip the directory and check the README file for instructions.)