fix "only photo responses" filter
Server side paging broke the photo responses filter in the responses tab because the filtering was being done on the front end after fetching. Now if a user asks to see only photo responses, the front end checks the campaign xml config to find out which surveys have photo prompts and queries for just those surveys.