Use sprites for small images
See client/resources/sprites/README.txt

Adds a ClientBundle resource with composite sprite images and css.
Renders small images (in menus, bullets, etc) with sprites to reduce number of round trips.

It was decided not to use sprites at this time, but the code is being checked in as
reference in case we decide to use them in the future