reduce data loads in admin tab
Previously, the user search api was queried for a list of every user in the system
every time the admin opened the Add New Members popup in the class edit tab. Now,
the list of users is loaded once, the first time the popup is opened, and is
only reloaded when a user is added or deleted.