Added class + user selection for mobility plots (issue #216)
Note: Class selection is a special case that is unique to Mobility plots. In order to determine which classes the current user is privileged, we use class/read with the class_role parameter to obtain only accessible classes. If no classes are available, the user should still have a "No class (self-only)" option to view one's own data.